requirements.txt
altgraph==0.17.4
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pefile==2023.2.7
pillow==12.0.0
py-ezmail==2.3.5
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pystray==0.19.5
pywin32-ctypes==0.2.3
setuptools==80.9.0
six==1.17.0
